Removed unneeded streaming operator function
This function causes clang with cxx11 features to build failures
It seems that this compiler setup evaluates the funtion at compiletime
of this lib. Nevertheless the whole function is not needed because
it postprones the error to here. Otherwise the user would get a warning
that his required specialization of the streaming is not implemented.