feat: implement SequentialReadDispatcher
This is a high-level class that allow someone to declare the various
streams that it "listens to" and apply said declaration(s) to a single
log file. The general application is a pipeline where test datasets
are being prepared, separately, from a general log dataset and can be
processed as-is (from a single file). The goal being that the test
is represented by said logfile, with no or little extra data.